reorganizing folder structure to highlight improvements
