+++ project/ddpg/plot.py
@@ -1,0 +1,11 @@
+import numpy as np
+import matplotlib.pyplot as plt
+
+#plots the scores of the course of training
+def plot_learning_curve(x, scores, figure_file):
+    running_avg = np.zeros(len(scores))
+    for i in range(len(running_avg)):
+        running_avg[i] = np.mean(scores[max(0, i-100):(i+1)])
+    plt.plot(x, running_avg)
+    plt.title('Running average of previous 100 scores')
+    plt.savefig(figure_file)
