+++ project/ddpg/time_converter.py
@@ -5,12 +5,10 @@ def timespan_format(timespan):
     m = (timespan % (60*60)) // 60
     s = timespan % 60
     time = f"{h:02}:{m:02}:{s:02}"
-    print(time)
     if h_full > 24:
         d = timespan // (60*60*24)
         h = (timespan % (60*60*24)) // (60*60)
-        time = f"{d:02}:{h:02}:{m:02}:{s:02}"
-    print(time)
+        time = f"{d:02}::{h:02}:{m:02}:{s:02}"
     return time
 
 timespan_format(303601)
@@ -20,5 +18,5 @@ if test:
     assert timespan_format(60*60) == "01:00:00"
     assert timespan_format(60*60+1) == "01:00:01"
     assert timespan_format(60*60+60) == "01:01:00"
-    assert timespan_format(24*60*60+1) == "01:00:00:01"
-    assert timespan_format(303601) == "03:12:20:01" # 3 days 12 hours and 20 minutes 1 second
+    assert timespan_format(24*60*60+1) == "01::00:00:01"
+    assert timespan_format(303601) == "03::12:20:01" # 3 days 12 hours 20 minutes 1 second
